main_app/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
SQLAlchemy==2.0.23
Werkzeug==3.0.1
python-dotenv==1.0.0 